+++ Gulpfile.js
@@ -127,7 +127,7 @@ export function build(done) {
           },
           { src: ["README.md"], dest: DESTINATION },
           { src: ["CHANGELOG.md"], dest: DESTINATION },
-          { src: ["flinksql-language-config.json"], dest: DESTINATION },
+          { src: ["flinksql-language-config.json", "flinksql.tmLanguage.json"], dest: DESTINATION },
         ],
       }),
       sentryRollupPlugin({