+++ src/models/organization.ts
@@ -1,7 +1,8 @@
 import { Data, type Require as Enforced } from "dataclass";
+import { OrganizationId } from "./resource";
 
 export class CCloudOrganization extends Data {
-  id!: Enforced<string>;
+  id!: Enforced<OrganizationId>;
   name!: Enforced<string>;
   current!: Enforced<boolean>;
   jit_enabled!: Enforced<boolean>;