+++ tests/unit/testResources/flinkStatement.ts
@@ -2,14 +2,11 @@ import { ConnectionType } from "../../../src/clients/sidecar";
 import { CCLOUD_CONNECTION_ID } from "../../../src/constants";
 import {
   FlinkStatement,
-  FlinkStatementMetadata,
-  FlinkStatementSpec,
-  FlinkStatementStatus,
-  FlinkStatementTraits,
 } from "../../../src/models/flinkStatement";
-import { EnvironmentId } from "../../../src/models/resource";
+import { EnvironmentId, OrganizationId } from "../../../src/models/resource";
 import { TEST_CCLOUD_ENVIRONMENT_ID } from "./environments";
 import { TEST_CCLOUD_FLINK_COMPUTE_POOL_ID } from "./flinkComputePool";
+import { TEST_CCLOUD_ORGANIZATION } from "./organization";
 
 export const TEST_CCLOUD_FLINK_STATEMENT = createFlinkStatement();
 
@@ -21,6 +18,7 @@ export interface CreateFlinkStatementArgs {
   sqlStatement?: string;
 
   environmentId?: EnvironmentId;
+  organizationId?: OrganizationId;
   computePoolId?: string;
 }
 
@@ -29,33 +27,35 @@ export function createFlinkStatement(overrides: CreateFlinkStatementArgs = {}): 
     connectionId: CCLOUD_CONNECTION_ID,
     connectionType: ConnectionType.Ccloud,
     environmentId: overrides.environmentId || TEST_CCLOUD_ENVIRONMENT_ID,
+    organizationId: overrides.organizationId || TEST_CCLOUD_ORGANIZATION.id,
     name: overrides.name || "statement0",
-    metadata: new FlinkStatementMetadata({
-      createdAt: new Date(),
-      updatedAt: new Date(),
-    }),
-    status: new FlinkStatementStatus({
+    metadata: {
+      self: null,
+      created_at: new Date(),
+      updated_at: new Date(),
+    },
+    status: {
       phase: overrides.phase || "RUNNING",
       detail: overrides.detail || "Running",
-      traits: new FlinkStatementTraits({
-        sqlKind: overrides.sqlKind || "SELECT",
-        bounded: true,
-        appendOnly: true,
+      traits: {
+        sql_kind: overrides.sqlKind || "SELECT",
+        is_bounded: true,
+        is_append_only: true,
         schema: {},
-      }),
-      scalingStatus: {},
-    }),
-    spec: new FlinkStatementSpec({
-      computePoolId: overrides.computePoolId || TEST_CCLOUD_FLINK_COMPUTE_POOL_ID,
-      sqlStatement: overrides.sqlStatement || "SELECT * FROM test_table",
+      },
+      scaling_status: {}
+    },
+    spec: {
+      compute_pool_id: overrides.computePoolId || TEST_CCLOUD_FLINK_COMPUTE_POOL_ID,
+      statement: overrides.sqlStatement || "SELECT * FROM test_table",
       principal: "test-principal",
-      authorizedPrincipals: [],
+      authorized_principals: [],
       properties: {
         "sql.current-catalog": "custom-data-env",
         "sql.current-database": "Custom Data Dedicated Replica",
         "sql.local-time-zone": "GMT-04:00",
       },
       stopped: false,
-    }),
+    },
   });
 }